tests: don't run teardown_btrfs for ceph-disk testing
if the test directory resides on a btrfs filesystem, test/ceph-disk.sh
currently calls teardown_btrfs to issue a btrfs subvolume delete request
for all second level child directories.

This should not be performed, as none of the ceph-disks tests create
btrfs subvolumes. The deletion attempts result in a number of errors,
e.g.:
    teardown_btrfs: 28: btrfs subvolume delete test-ceph-disk/osd/ceph-X
    ERROR: 'test-ceph-disk/osd/ceph-X' is not a subvolume

http://tracker.ceph.com/issues/11821 Fixes: #11821

Signed-off-by: David Disseldorp <ddiss@suse.de>