osd: set handle to next pg on pgls end of pg
Calculate which PG comes next in the bit-reversed
hash ordering, and set response.handle to the lowest
hash in that PG's range.

We do this OSD side rather than client side so that
the client doesn't have to wait for the OSD's map
epoch in order to correctly calculate the next hash
in the case of PG splitting.

Signed-off-by: John Spray <john.spray@redhat.com>