crush: shared files should use kernel-doc
This reverts mapper.c hunks of 830752a52841 ("doc: fix doxygen
warnings").  It is shared with the kernel and uses kernel-doc for
in-code documentation.  If doxygen warnings are an issue, we should
probably just exclude the 7 shared files in Doxyfile - we don't have
@file in these files or EXTRACT_ALL/EXTRACT_STATIC in Doxyfile, which
means these files are essentially skipped as there is no C++ stuff in
them.

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>