osd/PGBackend: set correct shard in objects_list_partial
We need to list objects within the current shard only.  We could get away
with being sloppy about shard previously when the ghobject_t sort order
was broken, but not in the new world.  Here, it is only necessary that we
capture all generations of the object to get the marker.

Signed-off-by: Sage Weil <sage@redhat.com>