Common: Do not use CEPH_PAGE_SIZE when appending buffers in Ceph
Whenever one grows a Buffer in Ceph, it is increased by the system
PAGE_SIZE.

Unfortunately, when one is running with a 64KB PAGE_SIZE, they can
witness quite a large memory spike under certain workloads, over what
can be observed when using a 4KB PAGE_SIZE.

This patch adjusts the Buffer append logic s.t. the buffer is expanded
in chunks of 4KB no matter what PAGE_SIZE is in operation. On systems
that run only with a 4KB PAGE_SIZE, this patch should not have any
noticeable effect. For architectures that can run with a 64KB PAGE_SIZE
(such as ARM and PowerPC), this patch should help reduce the size of
memory spikes under certain workloads.

With this patch applied on 64-bit ARM, the peak resident memory use by
ceph-mds running on 64KB pages is reduced from 10GB to 1GB. (One can
exacerbate a spike in memory use by unmounting a CephFS filesystem.)
The size of the virtual memory use by ceph-mds is reduced by a similar
magnitude.

Signed-off-by: Steve Capper <steve.capper@linaro.org>