suite: append flavor key only if kernel version is specified
There are three cases: a) don't install a kernel, b) install a distro
kernel or c) install a specified version.  Appending 'flavor: basic' to
kernel task config in the former two cases is wrong.  In case a)
appending flavor is not only bogus but also leads to bug: a { 'kernel':
{ 'flavor': 'basic' } } config confuses kernel task and it installs
a default version of the kernel (branch master) on all remotes when the
user's intent was to not install any kernel.  As for case b), distro
kernels are built by distros and don't have flavors.

Fix this by appending 'flavor: basic' to kernel task config only in
case c).

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>