common/RWLock: allow disabling read/write lock counts
For some workloads, it is worthwile to disable adjusting of internal nrlock
and nwlock atomic variables (which isn't super cheap and may lock CPUs for
longer than with regular mutexes). Added new argument (track_locks) to
RWLock constuctor (with default value of True) which enables or disables
adjusting these variables if needed.
(R|W)Locker helper classes are around 40% faster when tracking is disabled
and direct get_read()/get_write()+unlock() methods are up to 120% faster
when tracking is disabled.

Signed-off-by: Piotr Dałek <piotr.dalek@ts.fujitsu.com>