suite: don't try to check ceph packages against kernel flavor
The whole kernel_flavor check in schedule_suite() is completely wrong.
ceph packages and kernel packages are built on different gitbuilders
and so a set of ceph flavors is distinct from a set a kernel flavors
(they happen to intersect because both have a 'basic' flavor but that's
it).  Querying ceph gitbuilders with kernel flavor (which is what it
currently does - has_packages_for_distro() talks to ceph gitbuilders)
is useless and only worked because the only kernel we currently care
about and use in qa is 'basic' kernel.  Remove the for i in set(...)
construct and only check install task (i.e. ceph) flavor.

Also, crucially, do not make an empty dict a kernel task config in
!kernel_hash (i.e. "do not install any kernel") case.  This makes
kernel task install a default version of the kernel (branch master) on
all remotes.

Both of the above are partial reverts of commit 9968c366b17f ("Also
check install task's flavor when scheduling").

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>