mds: store layout on header object
This is surprisingly simple because we were already redundantly
calling store_backtrace whenever the layout changed!  That was
a side effect of the way add_old_pool is handled, the backtrace
version is bumped to latest even if the "old" pool is the
current one.

The upshot is that if we accept the existing behaviour of
also unnecessarily updating the 'parent' xattr, keeping
the new 'layout' xattr update requires no new dirty flags.  This
is a twitchy enough behaviour that new tests are needed to guard
against regressions though.

Fixes: #4161
Signed-off-by: John Spray <john.spray@redhat.com>