librbd: prevent object map updates from being interrupted
Object map updates were being canceled in-flight when the exclusive lock
is released.  This resulted in an ERESTART error code bubbling up to
AioRequest.

Fixes: 12165
Backport: hammer
Signed-off-by: Jason Dillaman <dillaman@redhat.com>
(cherry picked from commit 590cdc90edaf4f4ff06c97eb2f43b92ab9b60084)

Conflicts:
	src/librbd/ObjectMap.h
conflict due to a variable `m_snap_id' which was introduced in PR #4140
which is dropped as we are not backporting that feature