mon: MonOpRequest: allow setting the op type
When we introduced the MonOpRequest in the monitor and moved pretty much
every single function receiving messages in their arguments to take op
requests, we basically lost the type safety that was guaranteed from
Monitor::dispatch().

This patch adds an op_type field to the op request, as an easy fix for
this now lacking safety.

Signed-off-by: Joao Eduardo Luis <joao@suse.de>