rgw/rgw_main.cc:modified RGWFrontEndConfig
Modified RGWFrontEndConfig::parse_config function, which is responsible
for parsing rgw_frontends from ceph config file. Before modifciation
this would only work with frontends like this:
civetweb port=80
If we added a whitespace anywhere between 'port' and '80' it would
break, for example:
civetweb port = 80
The reason for this was dividing the above string with space being
a delimiter - 'port', '=' and '80' ending up as different strings
instead of being the same one(which was later being searched for '='
to find key value pair).
This has been changed in the way that now the first string up until
space is treated as a framework name  and the rest is being treated as
a key or key value pair.

Fixes: #12038

Signed-off-by: Michal Jarzabek <stiopa@gmail.com>