mon: MDSMonitor: keep last_metadata in memory
Previously, update_metadata() read last_metadata from the store, added
a new metadata to the transaction, and triggered propose, so metadata
eventually was updated in the store. But if during this time window
another update_metadata() happened, it read old metadata (without the
new mds from the first update), and when it was committed the new mds
from the first update was lost.

Signed-off-by: Mykola Golub <mgolub@mirantis.com>