get_latest_image_version_deb(): run apt-get update
When we were still using Chef, we accidentally ran it twice on downburst
VMs during scheduled jobs: once before the kernel task, and once after.
Now that we're using Ansible, and not running the one before, and
because the kernel task is special and runs before most tasks (even the
ansible task that might be part of the job), the kernel task might run
before apt has refreshed its cache. This is more likely to be a problem
on VMs, because their cache is certainly more stale, but could affect
physical hosts as well.

Make all deb-based systems update their cache before attempting to
install a kernel.

Signed-off-by: Zack Cerza <zack@redhat.com>