mds: s/MDS/MDSDaemon/ for clarity
It's now a bit confusing to have a class just called "MDS" --
rename it to MDSDaemon to make the distinction clear between
this and MDSRank.

It's easy because there are hardly any references to it now!  This
patch also clears up a few stray references that were no longer
needed in OSDMap and Messenger.

Signed-off-by: John Spray <john.spray@redhat.com>