Allow subclasses of task.Ansible to provide an inventory group name
Subclasses of task.Ansible can now provide an inventory group name to
assign all nodes to when creating a dynamic inventory.

This PR also makes ansible.CephLab use the group 'testnodes'. Without this,
teuthology can not run the testnodes.yml playbook against it's locked nodes
using a dynamic inventory as it's restricted to the testnodes group.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>