buffer.cc: add possibility to omit crc cache invalidation
In some cases, it might be worthwile to either not invalidate crc
cache at all, or postpone it until we're done with other tasks (for
example, after doing a lot of copy_in into single bufferptr instead
of invalid after each of copy_in). This reduces bufferptr lock usage
and in turn, decreases cpu utilization.
This change adds the optional crc_reset arg (which defaults to true)
so only new/aware code will use this feature.

Signed-off-by: Piotr Dałek <piotr.dalek@ts.fujitsu.com>