ObjectStore: simplify Sequencer API.
Sequencers were previously created directly by users of ObjectStore.
However, they were actually empty facades. The object was "completed"
once it was first passed to queue_transactions: an object deriving from
Sequencer_impl appropriate for the given store type was created
and attached to the Sequencer object via the "p" pointer.

This commit refactors the usage of Sequencer across the codebase.
Instead of splitting the implementation into two and "completing"
it when the object is first seen in queue_transactions, sequencers
are created using a new virtual method on ObjectStore, called
create_sequencer, and are already of the derived type. Changes are
also made to SharedPtrRegistry to accomodate the new creation pattern.

Signed-off-by: Krzysztof Kosiński <krzysztof.kosinski@intel.com>