tests: robust test for the pool create crushmap test
The test that goes with 4ea88bccf0a5db458bca3d204d195b05ec3eeffb to
verify that a bugous crush ruleset will prevent the creation of a pool
trying to use it was fragile. I depends on the implementation of the
erasure code lrc plugin and turns out to not work on i386.

The test is modified to use a fake crushtool that always returns false
and validate that it prevents the creation of a pool, which demonstrate
it is used for crushmap validation prior to the pool creation.

http://tracker.ceph.com/issues/12419 Fixes: #12419

Signed-off-by: Loic Dachary <ldachary@redhat.com>