rgw: fix radosgw start-up script.
radosgw init script is unable to start radosgw daemon.
as it is relies on requiretty being disabled.
once init script start this daemon with sudo it fails
to start the daemon.

changing 'sudo' to 'su', it will fix this issue and
will also help running radosgw daemon with our new
"ceph" UID project.

Fixes: #10927

Signed-off-by: Vikhyat Umrao <vumrao@redhat.com>