log: fix fd locking during reopen_log_file
We need to ensure there are no users of m_fd while we reopen the file.
Otherwise, a racing thread may open a new file with the same fd and log
data may get erroneously written to that fd instead.  This can cause data
corruption if that fd happens to be a user objects and writeable.

Fixes: #11586
Backport: hammer, giant, firefly
Signed-off-by: Sage Weil <sage@redhat.com>