osd/PG: do some new interval work on init() too
Some work we need to do on any new interval, *including* when the PG is
first created and there is no "interval change".  Factor out the (mostly
new) bits of start_peering_interval() that also need to happen on the
initial interval after the PG is first instantiated into on_new_interval().

Signed-off-by: Sage Weil <sage@redhat.com>