mon: allow a monitor out of monmap to join
if a monitor is not listed in monmap, it sends an MMonJoin to join
the existing cluster. and this message is handled by MonmapMonitor
which is a subclass of PaxosService, and `PaxosService::dispatch()`
asserts on unexpected op request type. but MMonJoin is marked as
an OP_TYPE_MONITOR request in `Monitor::dispatch_op()`, not is
not changed later before being passed to `PaxosService::dispatch()`.
so we'd better add monitor messages into this function's white list.

Signed-off-by: Kefu Chai <kchai@redhat.com>