hobject_t: decode future hobject_t::get_min() properly
The post-hammer wip-temp branch changed hobject_t::get_min() so that pool
is INT64_MIN instead of -1 and neglected to deal with the encoding compat
with older versions.  Compensate on hammer by mapping INT64_MIN to -1
locally.  See commit ff99af38df830e215359bfb8837bf310a2023a4d.

Note that this means upgrades from hammer to post-hammer *must* include
this fix prior to the upgrade.  This will need to be well-documented in the
release notes.

Master gets a similar fix so that they map our min value to the new
INT64_MIN one on decode.

Fixes: #12536 (for hammer)
Signed-off-by: Sage Weil <sage@redhat.com>