calamari_setup: handle RHCS 1.2 (missing /mnt/MON)
In RHCS 1.2 we don't have a /mnt/MON directory. The intention of
35c6363a1eaf5011e26be8561ffcd4e9324d13a2 was to handle this condition,
but in 1.2, the non-zero return code makes Teuthology fail the whole
test.

We don't want *Teuthology* itself to act on the return code here; we
simply want to know what it was and structure the rest of the test
accordingly.