rgw: implement s3 encoding-type for get bucket
This change introduces handling for the encoding-type request
parameter on the get bucket operation. An object key may contain
characters which are not supported in XML. Passing the value "url" for
the encoding-type parameter will cause the key to be urlencoded in the
response.

Fixes: #12735
Signed-off-by: Jeff Weber <jweber@cofront.net>
(cherry picked from commit 180ca7b56ba91a3995c76dac698aa4ee31d9a2ce)