os/FileJournal: For linux use logic block size as alignment for directio.
For directio, it limited by offset and size. At present, it use
CEPH_PAGE_SIZE(4k) as alignment to limit offset and size. But in linux,
in really it limited by logic size(at present is 512Byte).
Using 512B have two benefits:
a)decrease the pad size,especially for small write.
b)reduce the memcopy when tune aligment of buffer.

Signed-off-by: Jianpeng Ma <jianpeng.ma@intel.com>