osd/ReplicatedPG: fix filter lifecycle in do_osd_ops
If we take an error path in [N]PGLS (objects_list_partial error
or bad handle) and FAILOK is set we may exec a second filter op
that reallocates a new filter without freeing the old one.

Simplify by freeing only at the end of the function or when a new
filter is (re)allocated.

Signed-off-by: Sage Weil <sage@redhat.com>