xio: defer header encoding to remove serialization
A final ordering of messages must be established before transfer to
Accelio, but it is not necessary to serialize general encoding and
sequencing of xio_msg structures.

The main change here is to encode and checksum the general message in
the calling thread, but move sequence determination and setup of the
Accelio private header into the Accelio portal thread, which is a
single thread of control.

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>