buffer: change bufferlist to boost::intrusive::list
This makes bufferlist responsible for the lifecycle of the ptr it is
stringing together in the list.  There should be no real change in
performance with this change as we are still doing the same number of
allocations.  The only real difference is that we are handling those
allocations instead of leaving them to STL list<>.

Rename _buffers -> _ptrs to make sure we catch all users and update them
according to the new rules.

Signed-off-by: Sage Weil <sage@redhat.com>