Adds task to wrap the aver utility
Aver is a utility that allows to test high-level assertions over
benchmarking output. This task allows to specify assertions as part of a
teuthology task description (in YAML) and check the assertion (via aver)
on the output of a benchmark (e.g. rados).

The input to the task is a file in the job's archive path. Since there
are no packages available for aver, the task downloads the binary and
deletes it after the execution is done.