osd: set handle to start of next PG on pgnls end of pg
For the namespaced PGNLS op, set the handle to the start
of the next PG when we reach the end of the current PG.

We do this OSD side rather than client side so that
the client doesn't have to wait for the OSD's map
epoch in order to correctly calculate the next hash
in the case of PG splitting.

Only do this when the cluster is in bitwise mode to avoid
totally weird results when the client is in nibblewise
mode--the nibblewise sort order doesn't map contiguous
ranges of the hash to PGs.

Signed-off-by: Sage Weil <sage@redhat.com>