tests: ceph-disk tests may use system ceph-{mon,osd}
Allow ceph-disk.sh to run to test ceph as installed from packages.

When run from sources, ceph-disk.sh is expected to use the binaries from
the source tree. It is enough to prepend . to the PATH. There is no need
to prefix each binary with ./

The virtualenv is also only necessary when running from sources and
setting it up for ceph-detect-init is only done if in the source tree.

Signed-off-by: Loic Dachary <ldachary@redhat.com>