Adds radosbenchsweep task
This tasks executes a parameter sweep of radosbench by varying the size
of the pool (# of replicas), the number of osds for a pool, object size
and radosbench mode. Each configuration can also be repeated multiple
times.

The task also processes the output of each radosbench invocation to get
the summary and format it as CSV. The columns that are included in the
CSV file can be configured.