tests: drop docker-tests.sh root and /dev support
Now that ceph-disk.sh activation tests switched to the ceph-disk suite
as found in https://github.com/ceph/ceph-qa-suite/tree/master/suites/ceph-disk,
it does not need for root and /dev support from docker-tests.sh. There
currently is no other use case and since it's generally not a good idea
to run anything as root in a container anyway.

Signed-off-by: Loic Dachary <ldachary@redhat.com>