common/hobject.cc: explicitly typecast p.value_.get_int()
These two changes were necessary to get the pull request to pass "make check".
The "p.value_.get_int()" assignment was introduced by d4aedeb, which is a
backport of c05f895. The "ghobject_t::NO_SHARD" was introduced by b4a4b75,
which is a backport of dcf09eae.

Signed-off-by: Nathan Cutler <ncutler@suse.com>