distros: alias a-supported-distro.yaml to CentOS 7
Some jobs are distribution agnostic and it is meaningless to run them on
all supported distributions. For such jobs it is convient to require
than they run on a supported distribution but we don't care which. This
can be done by including a-supported-distro.yaml in the suite sub
directory instead of linking to distros/supported.

Signed-off-by: Loic Dachary <loic@dachary.org>