suite: add config.suite_verify_ceph_hash when no gitbuilder
The suite_verify_ceph_hash configuration option is added to disable the
gitbuilder package verifications.

If True, teuthology-suite verifies that a package matching the ceph
branch exists in the gitbuilder. If False, no verification is done and
teuthology-suite assumes the packages are either not necessary to run
the task or they are created on demand.

Signed-off-by: Loic Dachary <loic@dachary.org>