Use the uuid we've created the partition with
Simplify and be more accurate with finding the new partition
by using the uuid we created the partition with instead of guessing what
the next device name will be.

Fall back to the old way for distributions that don't include
udev >= 211

Add tests for lvm based osds. Don't run this test with udev < 211

Signed-off-by: Joe Julian <jjulian@io.com>