ceph-disk: - is translated into / with systemd
When running ceph-disk trigger /dev/dm-1 with systemd, the path name is
translated into /dev/dm/1 because of systemd escape rules. Explicitly
translate - into \x2d for systemd to preserve the -.

It would be better to use systemd-escape
http://www.freedesktop.org/software/systemd/man/systemd-escape.html
but it does not appear to be generally available on CentOS 7 and
probably other distributions.

http://tracker.ceph.com/issues/13174 Fixes: #13174

Signed-off-by: Loic Dachary <ldachary@redhat.com>