ceph-disk: guard partprobe with udevadm settle
Call udevadm settle before and after partprobe.

A side effect of partprobe is to remove partitions and add them again.
The first udevadm settle waits for ongoing udev events to complete, just
in case one of them rely on an existing partition on dev.

The second udevadm settle guarantees to the caller that all udev events
related to the partition table change have been processed, i.e. the
95-ceph-osd.rules actions and mode changes, group changes etc. are
complete.

Signed-off-by: Loic Dachary <ldachary@redhat.com>