mon: MonMap: change error code to avoid clashing with caller's
RadosClient::connect() calls a bunch of functions from MonClient, which
in turn relies on MonMap::build_initial(). Returning ENOENT on the
latter all the way up to RadosClient::connect() clashes with another
function from MonClient. There may be a better way to handle this, but
this is the simplest, EHOSTUNREACH seems moderately adequate and it we
get to differentiate between the two.

Signed-off-by: Joao Eduardo Luis <joao@suse.de>