rgw: modify documents and help infos' descriptions to the usage of option date when executing command "log show"
DOCUMENTATION #13235

This modification concerns documents and help infos. To the documents, we delete original incorrect instruction about the usage of the option date which induces user to type the
value of date in the format of "YYYY-MM-DD" when executing "log show" command, and we add the actual usage of "YYYY-MM-DD-hh". To the help infos, we add the NOTE to make user
notice that the usage of option date here is different with default's and of cource we give the right formatting.

Signed-off-by: Kongming Wu <wu.kongming@h3c.com>