buildpackages: flock must --close
The process run by flock must not inherit the file descriptor because
this will cause the lock to be held forever should the command survive
the call to flock. This is for instance the case for the ssh-agent.

Signed-off-by: Loic Dachary <loic@dachary.org>
(cherry picked from commit 204e27fecda583c312c35e905b6c864a5c8a1925)